// src/app/interfaces/notify.interface.ts
export interface Notify {
    INT_TRACKINGNOTIFICATIONID: number;
    INT_ENTITYID: number;
    INT_IDTRACKINGDATE: number;
    INT_WAYID: number;
    VCH_WAY: string;
    INT_IDGROUPDATE: number;
    VCH_GROUPNAME: string;
    VCH_ENGLISHDESCRIPTION: string;
    VCH_SPANISHDESCRIPTION: string;   
    VCH_NAMEDATE:string;
}